WIP - Feature - grouping delegate
This PR adds an optional delegate field to a grouping extension. The delegate is authorised to mint new assets into the group.

The main use case for this change is a program controlling either a mint (candy machine) or a conversion from another asset type.

* Added delegate field to grouping extension
* Updated group processor to check grouping delegate if present
* Updated create instruction to accept an optional grouping_delegate signer

Todo: Build changes into interface program - this isn't rebuilt when running programs:build

Also perhaps worth considering not reinitialising the data when the max_size is written as there are likely cases where this may need to be updated post creation.